update: 로그아웃
