+++ src/stores/auth/useAuthStore.ts
@@ -1,8 +1,17 @@
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
-import { AuthState } from "../../types/auth/authState";
 
 // 전역으로 관리할 유저 상태 Store 생성 및 저장
+
+// 유저의 로그인 상태 모델 (토큰 등 인증 정보만 저장)
+export type AuthState = {
+  isLoggedIn: boolean;
+  isNew: boolean;
+  accessToken: string | null; // 추후 수정 필요
+  setAccessToken: (accessToken: string) => void; // 추후 수정 필요
+  login: (accessToken: string, isNew: boolean) => void; // 추후 수정 필요
+  logout: () => void;
+};
 
 export const useAuthStore = create<AuthState>()(
   persist(
